jobIdOverride int > long

+++ src/ESFA.DC.Logging/Interfaces/ILogger.cs
@@ -9,7 +9,7 @@ namespace ESFA.DC.Logging.Interfaces
             string message,
             Exception exception = null,
             object[] parameters = null,
-            int jobIdOverride = -1,
+            long jobIdOverride = -1,
             [CallerMemberName] string callerMemberName = "",
             [CallerFilePath] string callerFilePath = "",
             [CallerLineNumber] int callerLineNumber = 0);
@@ -18,7 +18,7 @@ namespace ESFA.DC.Logging.Interfaces
             string message,
             Exception exception = null,
             object[] parameters = null,
-            int jobIdOverride = -1,
+            long jobIdOverride = -1,
             [CallerMemberName] string callerMemberName = "",
             [CallerFilePath] string callerFilePath = "",
             [CallerLineNumber] int callerLineNumber = 0);
@@ -26,7 +26,7 @@ namespace ESFA.DC.Logging.Interfaces
         void LogWarning(
             string message,
             object[] parameters = null,
-            int jobIdOverride = -1,
+            long jobIdOverride = -1,
             [CallerMemberName] string callerMemberName = "",
             [CallerFilePath] string callerFilePath = "",
             [CallerLineNumber] int callerLineNumber = 0);
@@ -34,7 +34,7 @@ namespace ESFA.DC.Logging.Interfaces
         void LogDebug(
             string message,
             object[] parameters = null,
-            int jobIdOverride = -1,
+            long jobIdOverride = -1,
             [CallerMemberName] string callerMemberName = "",
             [CallerFilePath] string callerFilePath = "",
             [CallerLineNumber] int callerLineNumber = 0);
@@ -42,7 +42,7 @@ namespace ESFA.DC.Logging.Interfaces
         void LogInfo(
             string message,
             object[] parameters = null,
-            int jobIdOverride = -1,
+            long jobIdOverride = -1,
             [CallerMemberName] string callerMemberName = "",
             [CallerFilePath] string callerFilePath = "",
             [CallerLineNumber] int callerLineNumber = 0);
@@ -50,7 +50,7 @@ namespace ESFA.DC.Logging.Interfaces
         void LogVerbose(
             string message,
             object[] parameters = null,
-            int jobIdOverride = -1,
+            long jobIdOverride = -1,
             [CallerMemberName] string callerMemberName = "",
             [CallerFilePath] string callerFilePath = "",
             [CallerLineNumber] int callerLineNumber = 0);
